% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{abbreviate_class}
\alias{abbreviate_class}
\title{Abbreviate a class}
\usage{
abbreviate_class(x)
}
\arguments{
\item{x}{a character vector of class labels.}
}
\description{
Abbreviate a class
}
\examples{
abbreviate_class(c("factor", "numeric", "integer"))
}
