% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-args.R
\name{drop_pkg_mask}
\alias{drop_pkg_mask}
\title{Drop the package mask}
\usage{
drop_pkg_mask(pkg_string)
}
\arguments{
\item{pkg_string}{The name of a function, which may contain package prefix
that is to be removed.}
}
\description{
Drop the package mask
}
\examples{
simplificar:::drop_pkg_mask("ggplot2::geom()")
}
\keyword{internal}
