% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-args.R
\name{drop_unnamed_dots}
\alias{drop_unnamed_dots}
\title{Extracts dots and returns named elements}
\usage{
drop_unnamed_dots(...)
}
\arguments{
\item{...}{Values passed to \code{...}.}
}
\description{
Otherwise, there is a downstream problem if functions use \code{...} and there
are unnamed arguments, they are matched by position.
}
\keyword{internal}
