% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{dummy_types}
\alias{dummy_types}
\title{A dummy dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 8 columns.}
\usage{
dummy_types
}
\description{
An excerpt of the mtcars dataset, renamed columns so for each type
(numeric, integer, character, factor), we have two columns.
}
\keyword{datasets}
