% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-system.R
\name{file_path}
\alias{file_path}
\title{A file.path that drops NULL}
\usage{
file_path(...)
}
\arguments{
\item{...}{Passed to \code{\link[base:file.path]{base::file.path()}}}
}
\description{
A file.path that drops NULL
}
\examples{
file_path(NULL, "hi")
}
