% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{generic_exported}
\alias{generic_exported}
\title{Generic data selection}
\arguments{
\item{data}{The data to use for plotting.}

\item{name}{The name of the variable to plot. If \code{NULL} (the default), the
name is inferred form \code{aes}.}

\item{names}{The names of the variables to plot. If \code{NULL} (the default),
the names are inferred form \code{aes}.}

\item{...}{Variables to select. You can use tidy selectors such as
\code{starts_with(...)}, see \code{\link[tidyselect:select_helpers]{tidyselect::select_helpers()}} for details.}

\item{geom}{The bare name of the geom to use. If \code{NULL} (the default),
it is determined by the class of the data to plot.}

\item{return_plot}{Whether or not the plot should, in addition to being
written to a file, be the return value of the function call.}
}
\description{
Generic data selection
}
\keyword{internal}
