% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{gg_conversions}
\alias{gg_conversions}
\alias{flatten_gg}
\alias{blow_gg}
\alias{pull_gg}
\title{Common conversions}
\usage{
flatten_gg(tbl)

blow_gg(...)

pull_gg(tbl, index = 1)
}
\arguments{
\item{tbl}{A tibble.}

\item{...}{Objects of class \code{gg}, that is, plots created with \code{ggplot}.}

\item{index}{An index. Positive numbers refer to the nth plot in the
gg table, negative numbers count backwards.}
}
\description{
Common conversation operations between objects of class gg (called raw gg)
and tabular extensions of it (called gg table) that contains meta
information.
}
\section{Functions}{
\itemize{
\item \code{flatten_gg}: Flattens a gg table into its raw counterpart.

\item \code{blow_gg}: Blows up a raw gg into a gg table.

\item \code{pull_gg}: extract a raw gg given it's row number in the
gg table.
}}

\examples{
gg_table <- vis_cols(mtcars, 2)
gg_raw <- gg_table \%>\%
  flatten_gg()
# the reverse operation (only column gg can be restored)
gg_raw \%>\%
  blow_gg()
}
