% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{merge_vis}
\alias{merge_vis}
\title{Merge visualizations into one}
\usage{
merge_vis(data, ...)
}
\arguments{
\item{data}{A \code{gg table}, that is, a data frame with plots, see \code{\link[=blow_gg]{blow_gg()}}.}

\item{...}{Further parameters passed to \code{\link[gridExtra:grid.arrange]{gridExtra::grid.arrange()}}.}
}
\description{
Simple wrapper around \code{\link[gridExtra:grid.arrange]{gridExtra::grid.arrange()}}.
}
\examples{
plots <- vis_cols(iris, transformer = vis_1d_distr)
merge_vis(plots)
}
\seealso{
Other post-creation manipulators: \code{\link{gg_conversions}},
  \code{\link{mutate_gg}}
}
