% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{dataset}
\name{mtcars_converted}
\alias{mtcars_converted}
\title{A transfered dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 11 columns.}
\usage{
mtcars_converted
}
\description{
A transfered dataset
}
\keyword{datasets}
