% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{mutate_gg}
\alias{mutate_gg}
\title{Mutate gg raw objects in a gg table}
\usage{
mutate_gg(data, add, ...)
}
\arguments{
\item{data}{A gg table.}

\item{add}{Terms to add to the plot.}

\item{...}{Positional indices that indicate the row for which the
corresponding raw gg object should be modified. If empty, all columns
are modified.}
}
\description{
A dplyr-like mutate to modify certain gg raw objects in a gg table.
}
\examples{
mtcars \%>\%
  vis_cols(vs, contains("hp"), "cyl", transformer = vis_2d_point) \%>\%
  mutate_gg(ggplot2::stat_summary(fun.y = mean, geom = "line"), 2, 3) \%>\%
  pull_gg(2)
}
\seealso{
Other post-creation manipulators: \code{\link{gg_conversions}},
  \code{\link{merge_vis}}
}
