% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefill.R
\name{prefill_aes}
\alias{prefill_aes}
\title{Prefill some aes arguments of a transformer}
\usage{
prefill_aes(transformer, ...)
}
\arguments{
\item{transformer}{A transformer function like \code{\link[=vis_2d_point]{vis_2d_point()}} that should
gain prefilled aesthetics.}

\item{...}{Named aesthetics, e.g. \code{color = "variable_x"}.}
}
\description{
Prefill some aes arguments of a transformer
}
\examples{
# If you prefill in-place, you need to supply `k_dimensional`, since it
# cannot be inferred from the transformer name itself.
mtcars \%>\%
  vis_cols(everything(),
    transformer = prefill_aes(vis_2d_point, color = "vs"), k_dimensional = 2
  ) \%>\%
  pull_gg(-1)
# Alternatively, you explificlty assign the object and give it a name
# that can be recognized by `vis_cols`.
vis_2d_point_with_color <- prefill_aes(vis_2d_point, color = "vs")
mtcars \%>\%
  vis_cols(everything(),
    transformer = vis_2d_point_with_color, k_dimensional = 2
) \%>\%
pull_gg(-1)
}
