% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-args.R
\name{set_dots}
\alias{set_dots}
\title{Extracts dots and returns named elements}
\usage{
set_dots(...)
}
\arguments{
\item{...}{Values passed to \code{...}.}
}
\description{
Otherwise, there is a downstream problem if functions use \code{...} and there
are unnamed arguments, they are matched by position.
}
