% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-args.R
\name{set_dots_distr}
\alias{set_dots_distr}
\title{Setting the dot argument depending on the geom used.}
\usage{
set_dots_distr(..., geom, fill)
}
\arguments{
\item{...}{Values to be passed to the geom.}

\item{geom}{The bare name of the geom to use.}

\item{fill}{The fill argument to the geom which may or may not be applicable.}
}
\description{
Setting the dot argument depending on the geom used.
}
