% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplificar-package.R
\docType{package}
\name{simplificar-package}
\alias{simplificar}
\alias{simplificar-package}
\title{simplificar: A high-level API for ggplot2}
\description{
simplificar - for simplify in Spanish - is a high-level
API for ggplot2 that enables fast experiments. It is envisioned as
a tool for exploratory analysis early in the cycle of statistical model
building where standardized plots should be available quickly, with few
lines of code.
}
\author{
\strong{Maintainer}: Lorenz Walthert \email{lorenz.walthert@icloud.com}

}
\keyword{internal}
