% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-system.R
\name{time_stamp}
\alias{time_stamp}
\title{Create a time stamp with a name}
\usage{
time_stamp(name = "x")
}
\arguments{
\item{name}{An identifier.}
}
\description{
Simple time stam generation.
}
\examples{
time_stamp("mtcars")
}
