% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{transform_cols}
\alias{transform_cols}
\title{Column transformer}
\usage{
transform_cols(data, cols, transformer, ...)
}
\arguments{
\item{data}{Data to convert.}

\item{cols}{Character vector with columns to convert.}

\item{transformer}{A quoted function name that applies the conversion,
e.g. \code{"as.numeric"}.}

\item{...}{Arguments passed to the \code{transformer}.}
}
\description{
Convert columns to a target class.
}
\examples{
mtcars \%>\%
  transform_cols("vs", "readr::parse_factor", levels = 0:1) \%>\%
  transform_cols("vs", "forcats::fct_recode", low = "0", high = "1")
}
