% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-args.R
\name{update_aes}
\alias{update_aes}
\title{Override aes using the geom and class for user convenience}
\usage{
update_aes(class, aes, geom)
}
\arguments{
\item{class}{The named vector with classes of the aesthetics.}

\item{aes}{The currentaesthetics itself.}

\item{geom}{The current geom.}
}
\description{
Certain geoms don't play nice with certain aesthetics, so the aesthetics are
modified to be more meaningful with that geom. See 'Examples'.
}
\examples{
library("ggplot2")
p <- ggplot(mtcars)
# nice
p + geom_violin(aes(factor(cyl), mpg))
# nonsense
p + geom_violin(aes(mpg, factor(cyl)))
mtcars \%>\%
  transform_cols("cyl", "as.factor") \%>\%
  vis_2d_distr(c("cyl", "mpg")) \%>\%
  pull_gg()
mtcars \%>\%
  transform_cols("cyl", "as.factor") \%>\%
  vis_2d_distr(c("mpg", "cyl")) \%>\%
  pull_gg()
}
