% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-vars.R
\name{vars_from_quo}
\alias{vars_from_quo}
\title{Select variables}
\usage{
vars_from_quo(quos, data, k_dimensional)
}
\description{
Wrapper around \code{\link[tidyselect:vars_select]{tidyselect::vars_select()}}.
}
