% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-1d-distr.R
\name{vis_1d_distr}
\alias{vis_1d_distr}
\alias{vis_1d_distr}
\alias{vis_1d_distr_to_file}
\title{Plot a distribution}
\usage{
vis_1d_distr(data, aes, name = aes, geom = NULL, fill = "gray60", ...)

vis_1d_distr_to_file(data, aes, name = aes, geom = NULL, fill = "gray60",
  sub_dir = deparse(substitute(data)), file = file_path(pkg_name(), sub_dir,
  time_stamp(name)), dimensions = rep(unit(5, "cm"), 2), device = "pdf",
  ...)
}
\arguments{
\item{data}{A data frame with data to visualize.}

\item{aes}{A string to indicate the column for which a distribution should
be shown.}

\item{name}{The name of the variable to plot. If \code{NULL} (the default), the
name is inferred form \code{aes}.}

\item{geom}{The bare name of the geom to use. If \code{NULL} (the default),
it is determined by the class of the data to plot.}

\item{fill}{The color to fill the density area with.}

\item{...}{Passed to the geom. See 'Details'.}

\item{sub_dir}{The sub-directory to save the plot. marely used to
construct the \code{file} argument.}

\item{file}{The path to the file where the visualization should be stored.}

\item{dimensions}{Height and width constructed with \code{\link[base:units]{base::units()}},
passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}

\item{device}{Passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\description{
Functions to create a distribution plot to be sent to the console or written
to a file.
}
\details{
The following types are implemented:
\itemize{
\item Histogram for discrete data with the following geom:
\code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}.
\item Density Plot for continuous data with the following geom:
\code{\link[ggplot2:geom_density]{ggplot2::geom_density()}}.
}

If the column has not the target class, you can convert one or multiple
columns easily with \code{\link[=transform_cols]{transform_cols()}}.
}
\section{Functions}{
\itemize{
\item \code{vis_1d_distr}: Create a visualization and returns it in the
console.

\item \code{vis_1d_distr_to_file}: The same as vis_1d_distr, but called for its side
effect, i.e. to save the visualization to a file.
}}

\examples{
vis_1d_distr(mtcars, aes = "cyl") \%>\%
  flatten_gg()
}
