% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-level-interface-point.R
\name{vis_2d_point}
\alias{vis_2d_point}
\alias{vis_2d_point_to_file}
\title{Create a scatter plot}
\usage{
vis_2d_point(data, aes, names = aes, geom = NULL, ...)

vis_2d_point_to_file(data, aes, name = aes,
  sub_dir = deparse(substitute(data)), file = file_path(pkg_name(), sub_dir,
  time_stamp(name)), dimensions = rep(unit(5, "cm"), 2), ...)
}
\arguments{
\item{data}{A data frame with data to visualize.}

\item{aes}{A character vector to indicate the columns that should be mapped.
Passed to \code{\link[ggplot2:aes_string]{ggplot2::aes_string()}}. Note that you can also plot aesthetics
other than x and y too.}

\item{names}{The names of the variables to plot. If \code{NULL} (the default),
the names are inferred form \code{aes}.}

\item{geom}{The bare name of the geom to use. If \code{NULL} (the default),
it is determined by the class of the data to plot.}

\item{...}{Passed to the geom. See 'Details'.}

\item{name}{The name of the variable to plot. If \code{NULL} (the default), the
name is inferred form \code{aes}.}

\item{sub_dir}{The sub-directory to save the plot. marely used to
construct the \code{file} argument.}

\item{file}{The path to the file where the visualization should be stored.
The file may contain an extension - or the extension can be specified with
the \code{device} argument.}

\item{dimensions}{Height and width constructed with \code{\link[grid:unit]{grid::unit()}},
passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\description{
Functions to create a scatter plot to be sent to the console or written
to a file.
}
\section{Functions}{
\itemize{
\item \code{vis_2d_point_to_file}: The same as vis_2d_point, but called for its side
effect, i.e. to save the visualization to a file.
}}

\examples{
# ordinary scatter plots for continuous data
plot <- vis_2d_point(iris,
  c(x = "Sepal.Width", y = "Sepal.Length", size = "Petal.Width")
)
plot \%>\%
  dplyr::filter(purrr::map_lgl(.data$aes, ~ any(.x \%in\% "Sepal.Length")))
# jitter plots if both variables are categorical
mtcars \%>\%
  transform_cols(c("vs", "am"), "as.factor") \%>\%
  vis_2d_point(c("vs", "am"), width = 0.1, height = 0.1) \%>\%
  pull_gg()
}
