% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{widh_heigh_from_aspect_ration}
\alias{widh_heigh_from_aspect_ration}
\title{Define the aspect ration and it's scale}
\usage{
widh_heigh_from_aspect_ration(ratio, scale = 5)
}
\arguments{
\item{ratio}{The aspect ratio.}

\item{scale}{How to transform the ration to obtain width and heigh.}
}
\description{
Convenient wrapper for specifying height and width.
}
