% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{width_heigh_from_aspect_ratio}
\alias{width_heigh_from_aspect_ratio}
\title{Define the aspect ration and it's scale}
\usage{
width_heigh_from_aspect_ratio(ratio = 16/9, scale = 5, units = "cm")
}
\arguments{
\item{ratio}{The aspect ratio.}

\item{scale}{How to transform the ration to obtain width and heigh.}

\item{units}{The measurement unit, e.g. "cm". See \code{\link[grid:units]{grid::units()}}.}
}
\description{
Convenient wrapper for specifying height and width.
}
