% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-vars.R
\name{vars_from_quo}
\alias{vars_from_quo}
\title{Select variables}
\usage{
vars_from_quo(quos, data_name, k_dimensional)
}
\arguments{
\item{quos}{Captured variables with \code{rlang::quos()}.}

\item{data_name}{All column names of the data from which to select.}

\item{k_dimensional}{The cardinality of the transformer.}
}
\description{
Wrapper around \code{\link[tidyselect:vars_select]{tidyselect::vars_select()}} with some pre-and post-processing.
Most notable, for any number of selected variables that are greater than
the number of dimensions, all sets of combinations of the variables with
cardinality \code{k_dimensional} are formed.
}
