% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{vis_to_file}
\alias{vis_to_file}
\title{Generic arguments}
\usage{
vis_to_file(transformer, data, aes, name = NULL, geom = NULL,
  sub_dir = NULL, file = file_path(pkg_name(), sub_dir, time_stamp(name)),
  dimensions = rep(unit(5, "cm"), 2), device = "pdf", ...)
}
\arguments{
\item{transformer}{\code{[template parameter]} The bare name of the tranformer to
use, e.g. \code{vis_2d_point}.}

\item{data}{The data to use for plotting.}

\item{aes}{A string to indicate the column for which a distribution should
be shown.}

\item{name}{The name of the variable to plot. If \code{NULL} (the default), the
name is inferred form \code{aes}.}

\item{geom}{The bare name of the geom to use. If \code{NULL} (the default),
it is determined by the class of the data to plot.}

\item{sub_dir}{The sub-directory to save the plot. marely used to
construct the \code{file} argument.}

\item{file}{The path to the file where the visualization should be stored.}

\item{dimensions}{Height and width constructed with \code{\link[base:units]{base::units()}},
passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}

\item{device}{Passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}

\item{...}{Variables to select. You can use tidy selectors such as
\code{starts_with(...)}, see \code{\link[tidyselect:select_helpers]{tidyselect::select_helpers()}} for details.}
}
\description{
Generic arguments
}
