% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_to_file.R
\name{vis_to_file}
\alias{vis_to_file}
\alias{vis_distr_to_file}
\alias{vis_1d_distr_to_file}
\alias{vis_2d_distr_to_file}
\alias{vis_point_to_file}
\alias{vis_1d_point_to_file}
\alias{vis_2d_point_to_file}
\title{Create a visualization and save it to a file}
\usage{
vis_distr_to_file(data, aes, name = aes, geom = NULL, fill = "gray60",
  sub_dir = deparse(substitute(data)), file = file_path(pkg_name(), sub_dir,
  time_stamp(name)), dimensions = rep(unit(5, "cm"), 2), device = "pdf",
  return_vis = FALSE, ...)

vis_1d_distr_to_file(data, aes, name = aes, geom = NULL, fill = "gray60",
  sub_dir = deparse(substitute(data)), file = file_path(pkg_name(), sub_dir,
  time_stamp(name)), dimensions = rep(unit(5, "cm"), 2), device = "pdf",
  return_vis = FALSE, ...)

vis_2d_distr_to_file(data, aes, name = aes, geom = NULL, fill = "gray60",
  sub_dir = deparse(substitute(data)), file = file_path(pkg_name(), sub_dir,
  time_stamp(name)), dimensions = rep(unit(5, "cm"), 2), device = "pdf",
  return_vis = FALSE, ...)

vis_point_to_file(data, aes, name = aes,
  sub_dir = deparse(substitute(data)), file = file_path(pkg_name(), sub_dir,
  time_stamp(name)), dimensions = rep(unit(5, "cm"), 2), device = "pdf",
  return_vis = FALSE, ...)

vis_1d_point_to_file(data, aes, name = aes,
  sub_dir = deparse(substitute(data)), file = file_path(pkg_name(), sub_dir,
  time_stamp(name)), dimensions = rep(unit(5, "cm"), 2), device = "pdf",
  return_vis = FALSE, ...)

vis_2d_point_to_file(data, aes, name = aes,
  sub_dir = deparse(substitute(data)), file = file_path(pkg_name(), sub_dir,
  time_stamp(name)), dimensions = rep(unit(5, "cm"), 2), device = "pdf",
  return_vis = FALSE, ...)
}
\arguments{
\item{data}{A data frame with data to visualize.}

\item{aes}{A string to indicate the column for which a distribution should
be shown.}

\item{name}{The name of the variable to plot. If \code{NULL} (the default), the
name is inferred form \code{aes}.}

\item{geom}{The bare name of the geom to use. If \code{NULL} (the default),
it is determined by the class of the data to plot.}

\item{fill}{The color to fill the density area with.}

\item{sub_dir}{The sub-directory to save the plot. merely used to
construct the \code{file} argument.}

\item{file}{The path to the file where the visualization should be stored.
The file may contain an extension - or the extension can be specified with
the \code{device} argument.}

\item{dimensions}{Height and width constructed with \code{\link[grid:unit]{grid::unit()}},
passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}

\item{device}{The device passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}

\item{return_vis}{Whether or not the plot should, in addition to being
written to a file, be the return value of the function call.}

\item{...}{Passed to the geom. See 'Details'.}
}
\value{
Returns the file name to which the visualization was written invisibly.
}
\description{
These functions do exactly what their counterparts without the \code{to_file}
extension in the function name do, but they save the output to a file.
They are called for their side effects only.
E.g. \code{\link[=vis_1d_distr_to_file]{vis_1d_distr_to_file()}} does what \code{\link[=vis_1d_distr]{vis_1d_distr()}}
does plus saving the output to a file.
}
